requirements.txt
numpy
scipy
matplotlib
sounddevice
soundfile
parselmouth
Pillow